organize: move remaining root files to proper locations
- Move demo scripts to scripts/ directory with documentation
- Archive legacy API files (api.py, health_app.py, prometheus_metrics_task.py) to docs/_archive/legacy/
- Create scripts/README.md documenting utility scripts usage
- Remove empty simple_bot.py file
- Clean root directory to only essential configuration files

This completes the comprehensive repository organization with:
- Clean root directory structure
- Proper categorization of all files
- Clear documentation for script usage
