--- run_phase25_tests.py
@@ -1,319 +1,0 @@
-"""
-🤖 Phase 2.5: AI/ML Enhancement Test Runner
-
-Test Coverage:
-- ML service initialization
-- Content analysis and optimization
-- Engagement prediction
-- Churn risk assessment
-- Performance analytics
-- API functionality
-"""
-
-import asyncio
-import logging
-import sys
-from datetime import datetime
-
-logging.basicConfig(level=logging.INFO, format="%(asctime)s - %(levelname)s - %(message)s")
-logger = logging.getLogger(__name__)
-
-
-async def test_ml_services():
-    """🧪 Test core ML services"""
-    print("\n" + "=" * 60)
-    print("🤖 PHASE 2.5: AI/ML ENHANCEMENT TESTING")
-    print("=" * 60)
-    test_results = {
-        "ml_imports": False,
-        "content_analysis": False,
-        "engagement_prediction": False,
-        "churn_prediction": False,
-        "service_integration": False,
-        "api_functionality": False,
-    }
-    print("\n🔍 Testing ML service imports...")
-    try:
-        from apps.bot.services.ml.churn_predictor import ChurnPredictor
-        from apps.bot.services.ml.content_optimizer import ContentOptimizer
-        from apps.bot.services.ml.engagement_analyzer import EngagementAnalyzer
-        from apps.bot.services.ml.prediction_service import ContentMetrics, PredictionService
-
-        print("✅ All ML services imported successfully")
-        test_results["ml_imports"] = True
-    except Exception as e:
-        print(f"❌ ML import failed: {e}")
-        return test_results
-    print("\n📝 Testing content analysis...")
-    try:
-        content_optimizer = ContentOptimizer()
-        test_content = "\n        🚀 Exciting news! Our new AI-powered analytics feature is now live!\n        \n        This revolutionary update will help you:\n        • Boost engagement by 40-60%\n        • Predict optimal posting times\n        • Analyze content performance in real-time\n        \n        Try it now and transform your social media strategy! \n        \n        #analytics #AI #socialmedia #growth #innovation\n        "
-        analysis = await content_optimizer.analyze_content(test_content)
-        print(f"   📊 Content Score: {analysis.overall_score:.1f}/100")
-        print(f"   😊 Sentiment: {analysis.sentiment_label} ({analysis.sentiment_score:.2f})")
-        print(f"   📖 Readability: {analysis.readability_level} ({analysis.readability_score:.1f})")
-        print(f"   📱 Hashtags: {analysis.hashtag_count}")
-        print(f"   💡 Suggestions: {len(analysis.optimization_tips)}")
-        if analysis.overall_score > 0:
-            print("✅ Content analysis working correctly")
-            test_results["content_analysis"] = True
-        else:
-            print("⚠️ Content analysis returned zero score")
-    except Exception as e:
-        print(f"❌ Content analysis failed: {e}")
-    print("\n⚡ Testing real-time content scoring...")
-    try:
-        scores = await content_optimizer.score_content_realtime(test_content)
-        print(f"   ⚡ Real-time Score: {scores.get('overall_score', 0):.2f}")
-        print(f"   📏 Length Score: {scores.get('length_score', 0):.2f}")
-        print(f"   📱 Hashtag Score: {scores.get('hashtag_score', 0):.2f}")
-        print(f"   😊 Sentiment Score: {scores.get('sentiment_score', 0):.2f}")
-        if scores.get("overall_score", 0) > 0:
-            print("✅ Real-time scoring working")
-    except Exception as e:
-        print(f"❌ Real-time scoring failed: {e}")
-    print("\n🎯 Testing engagement prediction...")
-    try:
-        prediction_service = PredictionService()
-        await prediction_service.initialize_models()
-        content_metrics = ContentMetrics(
-            sentiment_score=0.7,
-            readability_score=0.8,
-            hashtag_count=5,
-            word_count=150,
-            media_count=1,
-            emoji_count=3,
-            engagement_history=[120, 95, 180, 220, 150],
-        )
-        prediction = await prediction_service.predict_engagement(
-            content_metrics, channel_id=12345, scheduled_time=datetime.now()
-        )
-        print(f"   📈 Predicted Engagement: {prediction.prediction:.1f}")
-        print(f"   🎯 Confidence: {prediction.confidence:.1f}")
-        print(f"   🔑 Key Factors: {list(prediction.factors.keys())[:3]}")
-        print(f"   💡 Recommendations: {len(prediction.recommendations)}")
-        if prediction.prediction > 0:
-            print("✅ Engagement prediction working")
-            test_results["engagement_prediction"] = True
-    except Exception as e:
-        print(f"❌ Engagement prediction failed: {e}")
-    print("\n⏰ Testing optimal timing analysis...")
-    try:
-        timing_result = await prediction_service.find_optimal_posting_time(
-            channel_id=12345, content_type="general"
-        )
-        print(f"   ⏰ Optimal Times Found: {len(timing_result.get('optimal_times', {}))}")
-        print(f"   📈 Expected Boost: {timing_result.get('expected_boost', 1):.1f}x")
-        print(f"   🎯 Confidence: {timing_result.get('confidence', 0):.1f}")
-        monday_times = timing_result.get("optimal_times", {}).get("monday", [])
-        if monday_times:
-            print(f"   📅 Monday optimal times: {', '.join(monday_times[:3])}")
-        print("✅ Optimal timing analysis working")
-    except Exception as e:
-        print(f"❌ Optimal timing failed: {e}")
-    print("\n⚠️ Testing churn prediction...")
-    try:
-        churn_predictor = ChurnPredictor()
-        await churn_predictor.initialize_model()
-        assessment = await churn_predictor.predict_churn_risk(user_id=67890, channel_id=12345)
-        print(f"   ⚠️ Churn Probability: {assessment.churn_probability:.2f}")
-        print(f"   🚨 Risk Level: {assessment.risk_level}")
-        print(f"   🎯 Confidence: {assessment.confidence:.2f}")
-        print(f"   🔍 Risk Factors: {len(assessment.primary_risk_factors)}")
-        print(f"   💡 Retention Strategies: {len(assessment.retention_strategies)}")
-        print(f"   ⚡ Immediate Actions: {len(assessment.immediate_actions)}")
-        if assessment.churn_probability is not None:
-            print("✅ Churn prediction working")
-            test_results["churn_prediction"] = True
-    except Exception as e:
-        print(f"❌ Churn prediction failed: {e}")
-    print("\n📊 Testing engagement analyzer...")
-    try:
-        engagement_analyzer = EngagementAnalyzer(
-            prediction_service=prediction_service,
-            content_optimizer=content_optimizer,
-            churn_predictor=churn_predictor,
-        )
-        analysis = await engagement_analyzer.analyze_content_before_publishing(
-            content_text=test_content,
-            media_urls=["https://example.com/image.jpg"],
-            channel_id=12345,
-        )
-        print(f"   📊 Analysis Components: {len(analysis)}")
-        print(
-            f"   🎯 Publishing Score: {analysis.get('publishing_score', {}).get('overall_score', 0):.1f}"
-        )
-        print(
-            f"   📝 Content Score: {analysis.get('content_analysis', {}).get('overall_score', 0):.1f}"
-        )
-        print(f"   💡 Recommendations: {len(analysis.get('optimization_recommendations', []))}")
-        if "content_analysis" in analysis:
-            print("✅ Engagement analyzer working")
-            test_results["service_integration"] = True
-    except Exception as e:
-        print(f"❌ Engagement analyzer failed: {e}")
-    print("\n🏥 Testing service health checks...")
-    try:
-        health_checks = {}
-        services = {
-            "prediction_service": prediction_service,
-            "content_optimizer": content_optimizer,
-            "churn_predictor": churn_predictor,
-            "engagement_analyzer": engagement_analyzer,
-        }
-        for name, service in services.items():
-            if hasattr(service, "health_check"):
-                health = await service.health_check()
-                health_checks[name] = health
-                status = health.get("status", "unknown")
-                print(f"   {name}: {status}")
-        healthy_services = sum(1 for h in health_checks.values() if h.get("status") == "healthy")
-        print(f"✅ Health checks completed: {healthy_services}/{len(services)} services healthy")
-    except Exception as e:
-        print(f"❌ Health check failed: {e}")
-    print("\n📦 Testing container integration...")
-    try:
-        from apps.bot.container import OptimizedContainer
-
-        container = OptimizedContainer()
-        ml_providers = [
-            "prediction_service",
-            "content_optimizer",
-            "churn_predictor",
-            "engagement_analyzer",
-        ]
-        provider_count = 0
-        for provider in ml_providers:
-            if hasattr(container, provider):
-                provider_count += 1
-                print(f"   ✅ {provider} provider available")
-            else:
-                print(f"   ❌ {provider} provider missing")
-        if provider_count == len(ml_providers):
-            print("✅ All ML providers integrated in container")
-            test_results["service_integration"] = True
-        else:
-            print(f"⚠️ {provider_count}/{len(ml_providers)} ML providers integrated")
-    except Exception as e:
-        print(f"❌ Container integration test failed: {e}")
-    print("\n🌐 Testing API functionality...")
-    try:
-        api_request = {
-            "text": test_content,
-            "media_urls": ["https://example.com/image.jpg"],
-            "target_audience": "general",
-        }
-        content_analysis = await content_optimizer.analyze_content(
-            api_request["text"], api_request["media_urls"], api_request["target_audience"]
-        )
-        api_response = {
-            "overall_score": content_analysis.overall_score,
-            "sentiment": content_analysis.sentiment_label,
-            "optimization_tips": content_analysis.optimization_tips[:3],
-            "hashtag_suggestions": content_analysis.suggested_hashtags[:5],
-        }
-        print(f"   📊 API Response Generated: {len(api_response)} fields")
-        print(f"   🎯 Response Size: {len(str(api_response))} characters")
-        if api_response["overall_score"] > 0:
-            print("✅ API functionality working")
-            test_results["api_functionality"] = True
-    except Exception as e:
-        print(f"❌ API functionality test failed: {e}")
-    print("\n" + "=" * 60)
-    print("📊 PHASE 2.5 TEST RESULTS")
-    print("=" * 60)
-    passed_tests = sum(test_results.values())
-    total_tests = len(test_results)
-    success_rate = passed_tests / total_tests * 100
-    for test_name, result in test_results.items():
-        status = "✅ PASS" if result else "❌ FAIL"
-        print(f"   {test_name:<20}: {status}")
-    print(f"\n🎯 Overall Success Rate: {passed_tests}/{total_tests} ({success_rate:.1f}%)")
-    if success_rate >= 80:
-        print("🚀 EXCELLENT: Phase 2.5 AI/ML Enhancement is ready for deployment!")
-    elif success_rate >= 60:
-        print("✅ GOOD: Phase 2.5 core functionality working, minor issues detected")
-    elif success_rate >= 40:
-        print("⚠️ PARTIAL: Some Phase 2.5 components working, needs attention")
-    else:
-        print("❌ CRITICAL: Phase 2.5 needs significant fixes before deployment")
-    print("\n📈 EXPECTED PERFORMANCE IMPROVEMENTS:")
-    print("   • Content Engagement: +40-60% improvement")
-    print("   • User Retention: +20-30% increase")
-    print("   • Content Quality: +50-70% optimization")
-    print("   • Prediction Accuracy: >75% for engagement, >85% for churn")
-    return test_results
-
-
-async def test_dependencies():
-    """🔍 Test ML dependencies availability"""
-    print("\n📦 Testing ML dependencies...")
-    dependencies = ["scikit-learn", "numpy", "pandas", "textstat", "emoji", "joblib"]
-    available_deps = []
-    missing_deps = []
-    for dep in dependencies:
-        try:
-            if dep == "scikit-learn":
-                import sklearn
-
-                available_deps.append(dep)
-            elif dep == "numpy":
-                import numpy
-
-                available_deps.append(dep)
-            elif dep == "pandas":
-                import pandas
-
-                available_deps.append(dep)
-            elif dep == "textstat":
-                import textstat
-
-                available_deps.append(dep)
-            elif dep == "emoji":
-                import emoji
-
-                available_deps.append(dep)
-            elif dep == "joblib":
-                import joblib
-
-                available_deps.append(dep)
-        except ImportError:
-            missing_deps.append(dep)
-    print(f"✅ Available: {', '.join(available_deps)}")
-    if missing_deps:
-        print(f"❌ Missing: {', '.join(missing_deps)}")
-        print(f"💡 Install with: pip install {' '.join(missing_deps)}")
-    return len(missing_deps) == 0
-
-
-async def main():
-    """🚀 Main test runner"""
-    print("🤖 PHASE 2.5: AI/ML ENHANCEMENT - TEST SUITE")
-    print("=" * 60)
-    deps_ok = await test_dependencies()
-    if not deps_ok:
-        print("\n⚠️ Some dependencies missing. Install them first:")
-        print("pip install scikit-learn numpy pandas textstat emoji joblib")
-        return
-    results = await test_ml_services()
-    print(f"\n🏁 Testing completed at {datetime.now().strftime('%Y-%m-%d %H:%M:%S')}")
-    return results
-
-
-if __name__ == "__main__":
-    try:
-        results = asyncio.run(main())
-        passed = sum(results.values()) if results else 0
-        total = len(results) if results else 0
-        if passed == total:
-            sys.exit(0)
-        elif passed >= total * 0.8:
-            sys.exit(1)
-        else:
-            sys.exit(2)
-    except KeyboardInterrupt:
-        print("\n⚠️ Testing interrupted by user")
-        sys.exit(3)
-    except Exception as e:
-        print(f"\n❌ Testing failed with error: {e}")
-        sys.exit(4)
